feat(kgsdk): 属性分组

See merge request kgcloud/kg-cloud-kgms!268

+++ kgsdk/src/main/java/com/plantdata/kgcloud/plantdata/req/explore/common/AbstrackGraphParameter.java
@@ -27,7 +27,7 @@ public abstract class AbstrackGraphParameter {
     private List<String> allowAttsKey;
     private List<Long> allowTypes;
     private List<String> allowTypesKey;
-    private List<Integer> allowAttrGroups;
+    private List<Long> allowAttrGroups;
     private List<EntityScreeningBean> entityQuery;
     private GraphBean graphBean;
     private List<AttrScreeningBean> attAttFilters;
